Refactor configs: create akkeeper.configs; Introduce RichConfig and configs package object
